Added profile page, fi                                xed the tweets and its css, added images in tweets

+++ database/migrations/2024_05_29_134624_add_image_column_to_tweets.php
@@ -4,19 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddImageToTweetsTable extends Migration
+class AddImageColumnToTweets extends Migration
 {
     public function up()
     {
         Schema::table('tweets', function (Blueprint $table) {
-            $table->string('image')->nullable()->after('content'); // Adding the image column
+            $table->string('image')->nullable();
         });
     }
 
     public function down()
     {
         Schema::table('tweets', function (Blueprint $table) {
-            $table->dropColumn('image');
+            if (Schema::hasColumn('tweets', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
 }
